Use the app config lookup refactored from the app package
