+++ app/app_goxjs.go
@@ -50,10 +50,6 @@ func (a *fyneApp) SendNotification(n *fyne.Notification) {
 	}
 }
 
-func rootConfigDir() string {
-	return "/data/"
-}
-
 var themeChanged = js.FuncOf(func(this js.Value, args []js.Value) interface{} {
 	if len(args) > 0 && args[0].Type() == js.TypeObject {
 		fyne.CurrentApp().Settings().(*settings).setupTheme()
