+++ app/app_mobile_and.go
@@ -12,10 +12,7 @@ void sendNotification(uintptr_t java_vm, uintptr_t jni_env, uintptr_t ctx, char 
 */
 import "C"
 import (
-	"log"
 	"net/url"
-	"os"
-	"path/filepath"
 	"unsafe"
 
 	"fyne.io/fyne/v2"
@@ -43,14 +40,4 @@ func (a *fyneApp) SendNotification(n *fyne.Notification) {
 		C.sendNotification(C.uintptr_t(vm), C.uintptr_t(env), C.uintptr_t(ctx), titleStr, contentStr)
 		return nil
 	})
-}
-
-func rootConfigDir() string {
-	filesDir := os.Getenv("FILESDIR")
-	if filesDir == "" {
-		log.Println("FILESDIR env was not set by android native code")
-		return "/data/data" // probably won't work, but we can't make a better guess
-	}
-
-	return filepath.Join(filesDir, "fyne")
 }
