+++ app/app_mobile_ios.go
@@ -15,14 +15,8 @@ void sendNotification(char *title, char *content);
 import "C"
 import (
 	"net/url"
-	"path/filepath"
 	"unsafe"
 )
-
-func rootConfigDir() string {
-	root := C.documentsPath()
-	return filepath.Join(C.GoString(root), "fyne")
-}
 
 func (a *fyneApp) OpenURL(url *url.URL) error {
 	urlStr := C.CString(url.String())
