+++ internal/app/config.go
@@ -1,0 +1,5 @@
+package app
+
+func RootConfigDir() string {
+	return rootConfigDir()
+}
