+++ internal/app/config_desktop_darwin.go
@@ -1,0 +1,15 @@
+//go:build !ci && !ios && !wasm && !test_web_driver && !mobile
+
+package app
+
+import (
+	"os"
+	"path/filepath"
+)
+
+func rootConfigDir() string {
+	homeDir, _ := os.UserHomeDir()
+
+	desktopConfig := filepath.Join(filepath.Join(homeDir, "Library"), "Preferences")
+	return filepath.Join(desktopConfig, "fyne")
+}
