+++ internal/app/config_goxjs.go
@@ -1,0 +1,7 @@
+//go:build !ci && (!android || !ios || !mobile) && (wasm || test_web_driver)
+
+package app
+
+func rootConfigDir() string {
+	return "/data/"
+}
