+++ internal/app/config_mobile_ios.go
@@ -1,0 +1,8 @@
+//go:build !ci && ios && !mobile
+
+package app
+
+func rootConfigDir() string {
+	root := C.documentsPath()
+	return filepath.Join(C.GoString(root), "fyne")
+}
