+++ internal/app/config_windows.go
@@ -1,0 +1,10 @@
+//go:build !ci && !android && !ios && !wasm && !test_web_driver
+
+package app
+
+func rootConfigDir() string {
+	homeDir, _ := os.UserHomeDir()
+
+	desktopConfig := filepath.Join(filepath.Join(homeDir, "AppData"), "Roaming")
+	return filepath.Join(desktopConfig, "fyne")
+}
